scripts/guactober/requirements.txt
python-gitlab==4.10.0
PyGithub==2.4.0
gql==3.5.0
